requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
ta==0.10.2
pydantic==2.5.3
gunicorn==21.2.0
scipy==1.10.1